修改

+++ debian_begin.sh
@@ -4,7 +4,7 @@ cd /etc/apt
 debian_version=$(lsb_release -cs)
 start_step="${1:-restart}"
 
-if [ $start_step -eq "start" ];then
+if [ $start_step == "start" ];then
     mkfs.ext4 /dev/nvme0n1
     mkdir -p /mnt/nvme
     mount /dev/nvme0n1 /mnt/nvme
@@ -29,25 +29,27 @@ apt upgrade -y
 
 apt install -y ca-certificates curl software-properties-common
 
-#curl -fsSL https://mirrors.ustc.edu.cn/docker-ce/linux/debian/gpg | apt-key add -
-#mv trusted.gpg /etc/apt/trusted.gpg.d/docker-ce.gpg
-#add-apt-repository -y "deb [arch=arm64] https://mirrors.ustc.edu.cn/docker-ce/linux/debian ${debian_version} stable"
+if [ $start_step == "start" ];then
+    #curl -fsSL https://mirrors.ustc.edu.cn/docker-ce/linux/debian/gpg | apt-key add -
+    #mv trusted.gpg /etc/apt/trusted.gpg.d/docker-ce.gpg
+    #add-apt-repository -y "deb [arch=arm64] https://mirrors.ustc.edu.cn/docker-ce/linux/debian ${debian_version} stable"
 
-#curl -fsSL https://mirrors.ustc.edu.cn/docker-ce/linux/debian/gpg | gpg --import -
-#gpg --list-keys --with-colons #fpr字段为指纹
-#gpg --list-keys #key字段后面为id
-#KEYID_OR_FINGERPRINT=$(gpg --list-keys | sed -n '/Docker/{g;p;};h')
-#KEYID_OR_FINGERPRINT=$(gpg --list-keys | awk '/Docker/ {print prev} {prev=$0}')
-#gpg --output /etc/apt/keyrings/docker-ce.gpg --export $KEYID_OR_FINGERPRINT
-#echo "deb [arch=arm64 signed-by=/etc/apt/keyrings/docker-ce.gpg] https://mirrors.ustc.edu.cn/docker-ce/linux/debian ${debian_version} stable" > /etc/apt/sources.list.d/docker-ce.list
+    #curl -fsSL https://mirrors.ustc.edu.cn/docker-ce/linux/debian/gpg | gpg --import -
+    #gpg --list-keys --with-colons #fpr字段为指纹
+    #gpg --list-keys #key字段后面为id
+    #KEYID_OR_FINGERPRINT=$(gpg --list-keys | sed -n '/Docker/{g;p;};h')
+    #KEYID_OR_FINGERPRINT=$(gpg --list-keys | awk '/Docker/ {print prev} {prev=$0}')
+    #gpg --output /etc/apt/keyrings/docker-ce.gpg --export $KEYID_OR_FINGERPRINT
+    #echo "deb [arch=arm64 signed-by=/etc/apt/keyrings/docker-ce.gpg] https://mirrors.ustc.edu.cn/docker-ce/linux/debian ${debian_version} stable" > /etc/apt/sources.list.d/docker-ce.list
 
-rm -rf /etc/apt/keyrings/docker-ce.gpg
-curl -fsSL https://mirrors.ustc.edu.cn/docker-ce/linux/debian/gpg | gpg --dearmor -o /etc/apt/keyrings/docker-ce.gpg
-echo "deb [arch=arm64 signed-by=/etc/apt/keyrings/docker-ce.gpg] https://mirrors.ustc.edu.cn/docker-ce/linux/debian ${debian_version} stable" > /etc/apt/sources.list.d/docker-ce.list
+    rm -rf /etc/apt/keyrings/docker-ce.gpg
+    curl -fsSL https://mirrors.ustc.edu.cn/docker-ce/linux/debian/gpg | gpg --dearmor -o /etc/apt/keyrings/docker-ce.gpg
+    echo "deb [arch=arm64 signed-by=/etc/apt/keyrings/docker-ce.gpg] https://mirrors.ustc.edu.cn/docker-ce/linux/debian ${debian_version} stable" > /etc/apt/sources.list.d/docker-ce.list
+fi
 
 apt update
-
 apt install -y docker-ce
+
 mkdir -p /etc/docker
 mkdir -p /mnt/nvme/docker
 cat > /etc/docker/daemon.json << EOF
@@ -68,9 +70,12 @@ sudo systemctl restart docker
 
 usermod -aG docker zmh
 
-rm -rf /etc/apt/keyrings/kubernetes-apt-keyring.gpg
-curl -fsSL https://mirrors.ustc.edu.cn/kubernetes/core:/stable:/v1.31/deb/Release.key | gpg --dearmor -o /etc/apt/keyrings/kubernetes-apt-keyring.gpg
-echo "deb [arch=arm64 signed-by=/etc/apt/keyrings/kubernetes-apt-keyring.gpg] https://mirrors.ustc.edu.cn/kubernetes/core:/stable:/v1.31/deb/ /" > /etc/apt/sources.list.d/kubernetes.list
+if [ $start_step == "start" ];then
+    rm -rf /etc/apt/keyrings/kubernetes-apt-keyring.gpg
+    curl -fsSL https://mirrors.ustc.edu.cn/kubernetes/core:/stable:/v1.31/deb/Release.key | gpg --dearmor -o /etc/apt/keyrings/kubernetes-apt-keyring.gpg
+    echo "deb [arch=arm64 signed-by=/etc/apt/keyrings/kubernetes-apt-keyring.gpg] https://mirrors.ustc.edu.cn/kubernetes/core:/stable:/v1.31/deb/ /" > /etc/apt/sources.list.d/kubernetes.list
+fi
+
 apt update
 apt install -y kubelet kubeadm kubectl
 apt-mark hold kubelet kubeadm kubectl
@@ -78,11 +83,13 @@ apt-mark hold kubelet kubeadm kubectl
 #containerd config default > /etc/containerd/config.toml
 sed -i 's/disabled_plugins = ["false"]/disabled_plugins = ["cri"]/g'
 
-docker pull arm64v8/registry:latest
-docker run -d -p 5000:5000 --name registry -v /mnt/nvme/docker_registry:/var/lib/registry arm64v8/registry:latest
-docker pull arm64v8/debian:latest
-docker tag arm64v8/debian:latest localhost:5000/arm64v8/debian:latest
-docker push localhost:5000/arm64v8/debian:latest
+if [ $start_step == "start" ];then
+    docker pull arm64v8/registry:latest
+    docker run -d -p 5000:5000 --name registry -v /mnt/nvme/docker_registry:/var/lib/registry arm64v8/registry:latest
+    docker pull arm64v8/debian:latest
+    docker tag arm64v8/debian:latest localhost:5000/arm64v8/debian:latest
+    docker push localhost:5000/arm64v8/debian:latest
+fi
 
 apt install -y python3 python3-pip
 apt install -y vim
